Feature/fix file context (#13)
* feat:add cal function

* fix: use Date.now() for unique action ID in useActionCalculation

* feat: add common components and update imports for better organization
