+++ package.json
@@ -17,15 +17,13 @@
   "dependencies": {
     "@heroicons/react": "^2.2.0",
     "@tailwindcss/postcss": "^4.0.3",
-    "@xstate/react": "^5.0.2",
     "autoprefixer": "^10.4.20",
     "clsx": "^2.1.1",
     "mathjs": "^14.4.0",
     "postcss": "^8.5.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "tailwindcss": "^4.0.3",
-    "xstate": "^5.19.2"
+    "tailwindcss": "^4.0.3"
   },
   "devDependencies": {
     "@types/chrome": "^0.0.280",
