+++ src/components/widgets/common/Layout.tsx
