+++ src/components/widgets/common/SearchInput.tsx
