+++ src/entrypoints/popup/App.tsx
@@ -11,14 +11,14 @@ import usePopupShortcut from "@/hooks/usePopupShortcut";
 
 import Layout, {
   commonClassName as layoutClassName,
-} from "@/components/common/Layout";
+} from "@/components/widgets/common/Layout";
 import SearchInput, {
   commonClassName as searchInputClassName,
-} from "@/components/common/SearchInput";
+} from "@/components/widgets/common/SearchInput";
 import Badge from "@/components/modules/icon/Badge";
 import SquareBadge from "@/components/modules/icon/SquareBadge";
-import ResultFooter from "@/components/common/result/ResultFooter";
-import ResultLine from "@/components/common/result/item/ResultLine";
+import ResultFooter from "@/components/widgets/common/ResultFooter";
+import ResultLine from "@/components/widgets/common/ResultLine/ResultLine";
 
 import { ResultType } from "@/types/result";
